% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fav_fish.R
\name{fav_fish}
\alias{fav_fish}
\title{Print Name of Favourite Fish}
\usage{
fav_fish(SPC_Code)
}
\description{
This function takes an MNR species code and outputs a test string identifying your favourite fish.
}
